Changed z-slice behavior while tracking
During tracking, the GUI respects the current Z position. If there is a centroid for current cell in the specific z-slice that you are, the centroid will be red. If there there is a centroid for the cell in the timepoint, but in a different z-slice, it will be blue.

If tracking mode is inactive, but a cell is selected, the GUI changes Z-position while you explore the time series in order to keep the cell in focus. This is useful for evaluating tracking results.